+++ R/tidyABM-package.R
@@ -2,7 +2,7 @@
 "_PACKAGE"
 
 ## usethis namespace: start
-#' @importFrom rlang .data
+#' @importFrom rlang .data %||%
 #' @importFrom stats runif rnorm rlnorm setNames
 #' @importFrom utils combn
 ## usethis namespace: end